More generic rule for number, now accepts floats in different format (1.23 - 2E4 ) Grammar fix for negative and power Add a non-lua rule : a number with "heading ::plus::" like a = +4 (it doesn't parse in lua, but seems logic to implement)

+++ Visitor/Interpreter.php
@@ -272,7 +272,7 @@ class Interpreter implements \Hoa\Visitor\Visit {
                         return $value;
 
                     case 'number':
-                        return intval($value);
+                        return floatval($value);
 
                     case 'string':
                         return trim($value, '\'"');
